Revert "Refactor python"
