+++ python/configutil.py
