+++ python/genutil.py
